dashboards/api/requirements.txt
functions-framework==3.8.1
google-cloud-bigquery==3.25.0
